news_sense/backend/requirements.txt
fastapi
uvicorn
pymongo
transformers
torch
httpx
requests
beautifulsoup4
python-dotenv
google-generativeai

